+++ okuki-toothpick-mvp-sample/src/main/java/okuki/sample/kittens/detail/KittensDetailPlace.java
@@ -1,11 +1,11 @@
 package okuki.sample.kittens.detail;
 
+import okuki.Place;
 import okuki.PlaceConfig;
-import okuki.IntegerPlace;
 import okuki.sample.kittens.KittensPlace;
 
 @PlaceConfig(parent = KittensPlace.class)
-public class KittensDetailPlace extends IntegerPlace {
+public class KittensDetailPlace extends Place<Integer> {
 
     public KittensDetailPlace(Integer data) {
         super(data);