+++ okuki/src/main/java/okuki/SimplePlace.java
@@ -1,6 +1,6 @@
 package okuki;
 
-public class SimplePlace extends Place<Void> {
+public abstract class SimplePlace extends Place<Void> {
 
     public SimplePlace() {
         super(null);