+++ CHANGELOG.md
@@ -1,5 +1,8 @@
-## Release version 0.1.2 (Feb 11, 2017)
-* Bug fix for possible NPE in `Place.equals(obj)`
+## Release version 0.2.0 (Mar 09, 2017)
+* Toothpick integration refactoring
+* Toothpick auto-scoping
+* New Okuki-Android module
+* Okuki-Android Parcelable Okuki state save/restore
 
 ## Release version 0.1.1 (Feb 03, 2017)
 * Added generic typing to onPlace method to return Place without needing to cast.